+++ android/src/main/java/net/xpece/android/content/LongPreference.java
@@ -10,28 +10,28 @@ import android.content.SharedPreferences;
 public class LongPreference {
     private final SharedPreferences preferences;
     private final String key;
-    private final int defaultValue;
+    private final long defaultValue;
 
     public LongPreference(SharedPreferences preferences, String key) {
         this(preferences, key, 0);
     }
 
-    public LongPreference(SharedPreferences preferences, String key, int defaultValue) {
+    public LongPreference(SharedPreferences preferences, String key, long defaultValue) {
         this.preferences = preferences;
         this.key = key;
         this.defaultValue = defaultValue;
     }
 
-    public int get() {
-        return preferences.getInt(key, defaultValue);
+    public long get() {
+        return preferences.getLong(key, defaultValue);
     }
 
     public boolean isSet() {
         return preferences.contains(key);
     }
 
-    public void set(int value) {
-        preferences.edit().putInt(key, value).apply();
+    public void set(long value) {
+        preferences.edit().putLong(key, value).apply();
     }
 
     public void delete() {
